% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{any_null}
\alias{any_null}
\title{Convinience Helper to Test for any `NULL` value}
\usage{
any_null(x)
}
\arguments{
\item{x}{The test value.}
}
\value{
`TRUE` if any value in `x` is `NULL`.
}
\description{
Convinience Helper to Test for any `NULL` value
}
\seealso{
Other helper functions: \code{\link{any_na}},
  \code{\link{convert_values}}, \code{\link{has_dose}},
  \code{\link{has_units}}, \code{\link{is_dose_unit}},
  \code{\link{is_numeric_list}},
  \code{\link{is_single_logical}},
  \code{\link{is_single_numeric}},
  \code{\link{is_single_string}},
  \code{\link{is_string_list}}, \code{\link{is_time_unit}},
  \code{\link{is_unit}}, \code{\link{rm_units}},
  \code{\link{single_na}}, \code{\link{to_range}}
}
\concept{helper functions}
