% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{bind_master}
\alias{bind_master}
\title{Binds (Row Binds) two Master Objects}
\usage{
bind_master(master, slave)
}
\arguments{
\item{master}{The master object that meta data is preserved.}

\item{slave}{The slave master object that is bound to the master object.
The number of groups of the slave object must be at most the number of groups of the master
object. Missing group entries will be filled with `NA`.}
}
\value{
A bound (combined) master object.
}
\description{
Binds (Row Binds) two Master Objects
}
\examples{

mol_1 <- molecule("Foo", MW = 100)
mol_2 <- molecule("Bar", MW = 200)
a <- master("master", mol_1, groups = 2)
b <- master("slave", mol_2, groups = 1)
bind_master(a, b)
# bind_master(b, a) # fails

}
\seealso{
Other master functions: \code{\link{[.master}},
  \code{\link{add_master_entry}},
  \code{\link{entry.default}}, \code{\link{entry.master}},
  \code{\link{entry}}, \code{\link{filter_master}},
  \code{\link{groups.default}},
  \code{\link{groups.master}}, \code{\link{groups}},
  \code{\link{is_masterEntry_list}},
  \code{\link{is_masterEntry}}, \code{\link{is_master}},
  \code{\link{master}}, \code{\link{print.master}}
}
\concept{master functions}
