% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{convert_values}
\alias{convert_values}
\title{Converts Values from one unit into another.}
\usage{
convert_values(values, from = NA, to, attach.unit = F)
}
\arguments{
\item{values}{Input values that should be converted from `from` to `to`. Input can be a vector/`tibble`
or `data.frame`.}

\item{from}{Unit string that is compatible with `to`.}

\item{to}{f Unit string that is compatible with `from`.}

\item{attach.unit}{If `TRUE` the unit `from` is attached to the returned values.
If values already has a unit attached `from` is ignored.}
}
\value{
A
}
\description{
Converts Values from one unit into another.
}
\examples{

}
\seealso{
Other helper functions: \code{\link{any_na}},
  \code{\link{any_null}}, \code{\link{has_dose}},
  \code{\link{has_units}}, \code{\link{is_dose_unit}},
  \code{\link{is_numeric_list}},
  \code{\link{is_single_logical}},
  \code{\link{is_single_numeric}},
  \code{\link{is_single_string}},
  \code{\link{is_string_list}}, \code{\link{is_time_unit}},
  \code{\link{is_unit}}, \code{\link{rm_units}},
  \code{\link{single_na}}, \code{\link{to_range}}
}
\concept{helper functions}
