% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{entry}
\alias{entry}
\title{Generic Function to  Extract One or More entries (Row Entries) from an Object}
\usage{
entry(x, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list of entries.
}
\description{
Generic Function to  Extract One or More entries (Row Entries) from an Object
}
\seealso{
Other master functions: \code{\link{[.master}},
  \code{\link{add_master_entry}},
  \code{\link{bind_master}}, \code{\link{entry.default}},
  \code{\link{entry.master}}, \code{\link{filter_master}},
  \code{\link{groups.default}},
  \code{\link{groups.master}}, \code{\link{groups}},
  \code{\link{is_masterEntry_list}},
  \code{\link{is_masterEntry}}, \code{\link{is_master}},
  \code{\link{master}}, \code{\link{print.master}}
}
\concept{master functions}
