% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{excel_sheets}
\alias{excel_sheets}
\title{Title}
\usage{
excel_sheets(file)
}
\arguments{
\item{file}{}
}
\description{
Title
}
\seealso{
Other parser functions: \code{\link{extract_unit}},
  \code{\link{file_exists}}, \code{\link{file_type}},
  \code{\link{ident_time_column}},
  \code{\link{parse_master}}, \code{\link{read_file_table}}
}
\concept{parser functions}
