% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{extract_unit}
\alias{extract_unit}
\title{Extract a Unit from an Input String}
\usage{
extract_unit(str, extra.ident.fn = NULL)
}
\arguments{
\item{str}{That string that should contain the unit.}

\item{extra.ident.fn}{Extra identification function.

The input string `str` is searched for the pattern "[X]" with "X" being the unit.
If no pattern is found `NA` is returned. If the unit is unknown to `units` an error is raised.

The user can provide a function `extra.ident.fn` with the signature `function(x) = ...`.
If the unit cannot be identified `extra.ident.fn` is invoked (if not `NULL`).
The function should return a unit as `units::as_units(...)` or `stop` if it cannot be identified.
The pattern inside `[]` will be trimmed and converted to lower case.}
}
\value{
`NA` or the unit as `units::as_units`.
}
\description{
Extract a Unit from an Input String
}
\examples{

extract_unit("[DomInik]",
             function(x) if(x == "dominik") units::as_units("h") else stop("Oh noo"))

extract_unit("Bla bla [h]") # returns as_units("h")

}
\seealso{
Other parser functions: \code{\link{excel_sheets}},
  \code{\link{file_exists}}, \code{\link{file_type}},
  \code{\link{ident_time_column}},
  \code{\link{parse_master}}, \code{\link{read_file_table}}
}
\concept{parser functions}
