% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{fetch_properties_from_pubchem}
\alias{fetch_properties_from_pubchem}
\title{Fetch Compound properties from PubChem}
\usage{
fetch_properties_from_pubchem(names_or_cids,
  features = c("MolecularWeight"))
}
\arguments{
\item{names_or_cids}{Compound name or CID or vector/list-like compound names and/or CIDs}

\item{features}{The requested features from PubChem.
For a list of possbile features see \url{https://pubchemdocs.ncbi.nlm.nih.gov/pug-rest$_Toc494865567}.
If no features are provided (\code{NA} or \code{c()}) only the CIDs are returned.}
}
\value{
Dataframe with first column CID and requested feature columns.
}
\description{
Fetches properties like e.g. MW or XlogP3 from Pubchem for compounds or CIDs.
}
\examples{
nic_data <- fetch_from_pubchem("Nicotine")
mol_data <- fetch_from_pubchem(list("Nicotine", 12345), features = c("MolecularWeight", "XLogP"))

}
