% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{file_exists}
\alias{file_exists}
\title{Title}
\usage{
file_exists(file)
}
\arguments{
\item{file}{}
}
\description{
Title
}
\seealso{
Other parser functions: \code{\link{excel_sheets}},
  \code{\link{extract_unit}}, \code{\link{file_type}},
  \code{\link{ident_time_column}},
  \code{\link{parse_master}}, \code{\link{read_file_table}}
}
\concept{parser functions}
