% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{geo.dev.max}
\alias{geo.dev.max}
\title{Calculate the Maximum Geometric Deviation}
\usage{
geo.dev.max(values)
}
\arguments{
\item{values}{A numeric vector of values for which the maximum geometric deviation is to be calculated.}
}
\value{
A numeric value representing the maximum geometric deviation.
}
\description{
This function computes the maximum geometric deviation of a set of values.
It is calculated as the ratio of the geometric mean to the geometric standard deviation.
}
\examples{
values <- c(1, 2, 3, 4, 5)
geo.dev.max(values)
}
\seealso{
\code{\link{geo.mean}}, \code{\link{.geo.sd}}
}
