% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{geo.dev.min}
\alias{geo.dev.min}
\title{Calculate Geometric Deviation Minimum}
\usage{
geo.dev.min(values)
}
\arguments{
\item{values}{A numeric vector of values for which the geometric deviation minimum
is to be calculated. All values must be positive.}
}
\value{
A numeric value representing the geometric deviation minimum.
}
\description{
This function calculates the geometric deviation minimum for a given set of values.
It is computed as the product of the geometric mean of the values and the geometric
standard deviation of the values.
}
\examples{
values <- c(1, 2, 3, 4, 5)
geo.dev.min(values)

}
\seealso{
\code{\link{geo.mean}}, \code{\link{.geo.sd}}
}
