% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{geo.mean}
\alias{geo.mean}
\title{Calculate the Geometric Mean}
\usage{
geo.mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector containing the values for which the geometric mean is to be calculated.}

\item{na.rm}{A logical value indicating whether `NA` values should be removed before computation. Defaults to `TRUE`.}
}
\value{
A numeric value representing the geometric mean of the input vector.
}
\description{
This function computes the geometric mean of a numeric vector.
}
\examples{
geo.mean(c(1, 2, 3, 4, 5))
geo.mean(c(1, 2, NA, 4, 5), na.rm = TRUE)

}
