% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{groups.default}
\alias{groups.default}
\title{Default Function for Generic `groups`.}
\usage{
\method{groups}{default}(x, ...)
}
\arguments{
\item{x}{The object that contains or is associdated with groups.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Will raise an error (`stop`).
}
\description{
Default Function for Generic `groups`.
}
\seealso{
Other master functions: \code{\link{[.master}},
  \code{\link{add_master_entry}},
  \code{\link{bind_master}}, \code{\link{entry.default}},
  \code{\link{entry.master}}, \code{\link{entry}},
  \code{\link{filter_master}}, \code{\link{groups.master}},
  \code{\link{groups}}, \code{\link{is_masterEntry_list}},
  \code{\link{is_masterEntry}}, \code{\link{is_master}},
  \code{\link{master}}, \code{\link{print.master}}
}
\concept{master functions}
