% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{groups.master}
\alias{groups.master}
\title{Fetches the Groups defined in the Master Object}
\usage{
\method{groups}{master}(x, ...)
}
\arguments{
\item{x}{The master object}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A vector of group names or `NULL` if no groups were defined.
}
\description{
Fetches the Groups defined in the Master Object
}
\examples{

mol <- molecule("Mol 1", MW = 100)
m1 <- master("test", mol)
m2 <- master("test 2", mol, "my ref", groups = 3)
groups(m1) # returns NULL
groups(m2) # returns c("Group_1", "Group_2", "Group_3")

}
\seealso{
Other master functions: \code{\link{[.master}},
  \code{\link{add_master_entry}},
  \code{\link{bind_master}}, \code{\link{entry.default}},
  \code{\link{entry.master}}, \code{\link{entry}},
  \code{\link{filter_master}},
  \code{\link{groups.default}}, \code{\link{groups}},
  \code{\link{is_masterEntry_list}},
  \code{\link{is_masterEntry}}, \code{\link{is_master}},
  \code{\link{master}}, \code{\link{print.master}}
}
\concept{master functions}
