% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{groups}
\alias{groups}
\title{Generic Function to Return Groups of an Object}
\usage{
groups(x, ...)
}
\arguments{
\item{x}{The object that contains or is associdated with groups.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A list or vector of group names. Will return `NULL` if no groups are associated with
  the object `x`.
}
\description{
Generic Function to Return Groups of an Object
}
\seealso{
Other master functions: \code{\link{[.master}},
  \code{\link{add_master_entry}},
  \code{\link{bind_master}}, \code{\link{entry.default}},
  \code{\link{entry.master}}, \code{\link{entry}},
  \code{\link{filter_master}},
  \code{\link{groups.default}},
  \code{\link{groups.master}},
  \code{\link{is_masterEntry_list}},
  \code{\link{is_masterEntry}}, \code{\link{is_master}},
  \code{\link{master}}, \code{\link{print.master}}
}
\concept{master functions}
