% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{has_dose}
\alias{has_dose}
\title{Tests if a Single Numeric Value has a valid Dosing Unit attached}
\usage{
has_dose(value)
}
\arguments{
\item{value}{A single numeric value.}
}
\value{
`TRUE` if it is a single numeric value with a valid dosing unit (e.g. mg, µg, ...)
  attached, else `FALSE`.
}
\description{
Tests if a Single Numeric Value has a valid Dosing Unit attached
}
\examples{

has_dose(units::as_units(12, "mg")) # true
has_dose(units::as_units(12.4, "\\U00B5g")) # true
has_dose(units::as_units(12.4, "m")) # false
has_dose(12.4) # false

}
\seealso{
Other helper functions: \code{\link{any_na}},
  \code{\link{any_null}}, \code{\link{convert_values}},
  \code{\link{has_units}}, \code{\link{is_dose_unit}},
  \code{\link{is_numeric_list}},
  \code{\link{is_single_logical}},
  \code{\link{is_single_numeric}},
  \code{\link{is_single_string}},
  \code{\link{is_string_list}}, \code{\link{is_time_unit}},
  \code{\link{is_unit}}, \code{\link{rm_units}},
  \code{\link{single_na}}, \code{\link{to_range}}
}
\concept{helper functions}
