% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{has_fraction}
\alias{has_fraction}
\title{Tests if a matched profiles has molecules with fraction unit}
\usage{
has_fraction(profile)
}
\arguments{
\item{profile}{A matched profile}
}
\value{
TRUE if the molecule of any profile has fraction unit else FALSE
}
\description{
Tests if a matched profiles has molecules with fraction unit
}
