% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{has_molecules}
\alias{has_molecules}
\title{Checks if a Molecule List Contains all Query Ids}
\usage{
has_molecules(mol.list, query.ids)
}
\arguments{
\item{mol.list}{A non-empty molecule list.}

\item{query.ids}{One or more ids that should be queried}
}
\value{
True if all query ids could be found, else False.
}
\description{
Checks if a Molecule List Contains all Query Ids
}
\examples{

mol_1 <- molecule("Mol 1", MW = 100)
mol_2 <- molecule("Mol 2", MW = 200)
mol_list <- molecule_list(mol_1, mol_2)
has_molecules(mol_list, c("Mol 1")) # TRUE
has_molecules(mol_list, c("Mol 1", "Mol 2")) # TRUE
has_molecules(mol_list, c("Mol 3", "Mol 2")) # FALSE

}
\seealso{
Other molecule functions: \code{\link{is_molecule_list}},
  \code{\link{is_molecule}},
  \code{\link{is_unique_molecule_list}},
  \code{\link{merge_molecule_lists}},
  \code{\link{molecule_from_ids}},
  \code{\link{molecule_ids}}, \code{\link{molecule_list}},
  \code{\link{molecule}}, \code{\link{print.molecule}}
}
\concept{molecule functions}
