% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{has_units}
\alias{has_units}
\title{Tests if a Data Structure has Units Attached.}
\usage{
has_units(input)
}
\arguments{
\item{input}{The data structure that should be tested.}
}
\value{
True if the data structure inherits from \code{units} (must use the \code{units} package).
}
\description{
Tests if a Data Structure has Units Attached.
}
\seealso{
Other helper functions: \code{\link{is_single_logical}},
  \code{\link{is_single_numeric}},
  \code{\link{is_single_string}},
  \code{\link{is_string_list}}, \code{\link{is_unit}}
}
\concept{helper functions}
