% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{ident_column}
\alias{ident_column}
\title{Identifies and Returns a Match in a `tibble`/`data.frame` Header or String Vector}
\usage{
ident_column(x, match, multi.matches = F)
}
\arguments{
\item{x}{`tibble`/`data.frame` or String Vector}

\item{match}{A single string that should be identified}

\item{multi.matches}{If `TRUE` multiple matches are returned and not only the best match.}
}
\value{
The position in the string vector or header or `NA` is not found or ambiguous.
}
\description{
Identifies and Returns a Match in a `tibble`/`data.frame` Header or String Vector
}
\concept{parser functions

The function tries to match as follows:
  1. global match (has the match pattern)
  2. If ambiguous: Match strings that start with `match`
  3. If ambiguous: Match strings that start with `match` and have whitespace the match.

If `multi.matches` is set to `TRUE` only option (1) will be applied.}
