% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{ident_time_column}
\alias{ident_time_column}
\title{Tries to Identify a Time Entry in a `tibble`/`data.frame` Header or String Vector}
\usage{
ident_time_column(df, matches = c("time", "tad", "t"))
}
\arguments{
\item{df}{A `tibble`/`data.frame` or string vector.}

\item{matches}{String or vector of strings that are matched one after another.

The function tries to match all patterns defined in `matches` first.
  If no entry is found or ambiguous matches it searches for units ("[X]") that are convertible
  to a valid time unit.}
}
\value{
NA if no non-ambiguous entry could be identified, else the entry index.
}
\description{
Tries to Identify a Time Entry in a `tibble`/`data.frame` Header or String Vector
}
\seealso{
Other parser functions: \code{\link{excel_sheets}},
  \code{\link{extract_unit}}, \code{\link{file_exists}},
  \code{\link{file_type}}, \code{\link{parse_master}},
  \code{\link{read_file_table}}
}
\concept{parser functions}
