% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{is.molecule}
\alias{is.molecule}
\title{Check if an Object is a Molecule}
\usage{
is.molecule(x)
}
\arguments{
\item{x}{An object to be checked.}
}
\value{
A logical value: `TRUE` if the object inherits from the class "molecule", otherwise `FALSE`.
}
\description{
This function checks whether the given object inherits from the class "molecule".
}
