% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_dose_unit}
\alias{is_dose_unit}
\title{Tests for a Valid Dose Unit String}
\usage{
is_dose_unit(str)
}
\arguments{
\item{str}{The string that should be tested.}
}
\value{
`TRUE`, if a valid dose unit is provided, else `FALSE`.
}
\description{
This function will only work correctly for lower case strings (e.g. kg and not KG).
For non-strings, NA or NULL this function will return False.
}
\examples{

is_dose_unit("") # F
is_dose_unit(NA) # F
is_dose_unit("mg") # T
is_dose_unit("pg") # T
is_dose_unit("µg") # T
is_dose_unit("µg/kg") # F

}
\seealso{
Other helper functions: \code{\link{any_na}},
  \code{\link{any_null}}, \code{\link{convert_values}},
  \code{\link{has_dose}}, \code{\link{has_units}},
  \code{\link{is_numeric_list}},
  \code{\link{is_single_logical}},
  \code{\link{is_single_numeric}},
  \code{\link{is_single_string}},
  \code{\link{is_string_list}}, \code{\link{is_time_unit}},
  \code{\link{is_unit}}, \code{\link{rm_units}},
  \code{\link{single_na}}, \code{\link{to_range}}
}
\concept{helper functions}
