% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{is_master}
\alias{is_master}
\title{Checks if a Data Structure is a Master Object}
\usage{
is_master(ds)
}
\arguments{
\item{ds}{The data structure that should be tested.}
}
\value{
True if the data structure is a Master object
}
\description{
Checks if a Data Structure is a Master Object
}
\seealso{
Other master functions: \code{\link{add_master_entry}},
  \code{\link{bind_master}}, \code{\link{filter_master}},
  \code{\link{master}}, \code{\link{print.master}}
}
\concept{master functions}
