% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{is_masterEntry}
\alias{is_masterEntry}
\title{Checks if a Data Structure is a MasterEntry Object}
\usage{
is_masterEntry(x)
}
\arguments{
\item{x}{The data structure that should be tested.}
}
\value{
True if the data structure is a MasterEntry Object
}
\description{
Checks if a Data Structure is a MasterEntry Object
}
\seealso{
Other master functions: \code{\link{[.master}},
  \code{\link{add_master_entry}},
  \code{\link{bind_master}}, \code{\link{entry.default}},
  \code{\link{entry.master}}, \code{\link{entry}},
  \code{\link{filter_master}},
  \code{\link{groups.default}},
  \code{\link{groups.master}}, \code{\link{groups}},
  \code{\link{is_masterEntry_list}},
  \code{\link{is_master}}, \code{\link{master}},
  \code{\link{print.master}}
}
\concept{master functions}
