% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{is_masterEntry_list}
\alias{is_masterEntry_list}
\title{Checks if a Data Structure is a list of MasterEntry Objects}
\usage{
is_masterEntry_list(list)
}
\arguments{
\item{list}{The data structure that should be tested.}
}
\value{
`TRUE` if the data structure is a MasterEntry list. For single MasterEntry
  this function returns `FALSE`.
}
\description{
Checks if a Data Structure is a list of MasterEntry Objects
}
\seealso{
Other master functions: \code{\link{[.master}},
  \code{\link{add_master_entry}},
  \code{\link{bind_master}}, \code{\link{entry.default}},
  \code{\link{entry.master}}, \code{\link{entry}},
  \code{\link{filter_master}},
  \code{\link{groups.default}},
  \code{\link{groups.master}}, \code{\link{groups}},
  \code{\link{is_masterEntry}}, \code{\link{is_master}},
  \code{\link{master}}, \code{\link{print.master}}
}
\concept{master functions}
