% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{is_molecule}
\alias{is_molecule}
\title{Checks if a Data Structure is a Molecule}
\usage{
is_molecule(ds)
}
\arguments{
\item{ds}{The data structure that should be tested.}
}
\value{
True if the data structure is a Molecule.
}
\description{
Checks if a Data Structure is a Molecule
}
\seealso{
Other molecule functions: \code{\link{has_molecules}},
  \code{\link{is_molecule_list}},
  \code{\link{is_unique_molecule_list}},
  \code{\link{merge_molecule_lists}},
  \code{\link{molecule_from_ids}},
  \code{\link{molecule_ids}}, \code{\link{molecule_list}},
  \code{\link{molecule}}, \code{\link{print.molecule}}
}
\concept{molecule functions}
