% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{is_molecule_list}
\alias{is_molecule_list}
\title{Checks if a Data Structure is a list of Molecules}
\usage{
is_molecule_list(list)
}
\arguments{
\item{list}{The data structure that should be tested.}
}
\value{
True if the data structure is a Molecule list. For single Molecules this function returns False.
}
\description{
Checks if a Data Structure is a list of Molecules
}
\seealso{
Other molecule functions: \code{\link{has_molecules}},
  \code{\link{is_molecule}},
  \code{\link{is_unique_molecule_list}},
  \code{\link{match_mol}}, \code{\link{match_pvb}},
  \code{\link{merge_molecule_lists}},
  \code{\link{molecule_from_ids}},
  \code{\link{molecule_ids}}, \code{\link{molecule_list}},
  \code{\link{molecule}}, \code{\link{print.molecule}}
}
\concept{molecule functions}
