% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_numeric_list}
\alias{is_numeric_list}
\title{Tests if Input is a List-like Object of Numerics}
\usage{
is_numeric_list(vec, can.be.empty = F)
}
\arguments{
\item{vec}{The input object.}

\item{can.be.empty}{If NA-values are allowed.

Returns `TRUE` for empty lists.}
}
\value{
`TRUE`, if the input is a list-like object of numerics, else `FALSE`.
}
\description{
Tests if Input is a List-like Object of Numerics
}
\seealso{
Other helper functions: \code{\link{any_na}},
  \code{\link{any_null}}, \code{\link{convert_values}},
  \code{\link{has_dose}}, \code{\link{has_units}},
  \code{\link{is_dose_unit}},
  \code{\link{is_single_logical}},
  \code{\link{is_single_numeric}},
  \code{\link{is_single_string}},
  \code{\link{is_string_list}}, \code{\link{is_time_unit}},
  \code{\link{is_unit}}, \code{\link{rm_units}},
  \code{\link{single_na}}, \code{\link{to_range}}
}
\concept{helper functions}
