% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_single_logical}
\alias{is_single_logical}
\title{Checks if Input is a Single Logical Value}
\usage{
is_single_logical(input)
}
\arguments{
\item{input}{The input that should be tested}
}
\value{
`TRUE` if the input is a single logical value. Lists with one logical value will return `FALSE`.
}
\description{
Checks if Input is a Single Logical Value
}
\seealso{
Other helper functions: \code{\link{any_na}},
  \code{\link{any_null}}, \code{\link{convert_values}},
  \code{\link{has_dose}}, \code{\link{has_units}},
  \code{\link{is_dose_unit}},
  \code{\link{is_numeric_list}},
  \code{\link{is_single_numeric}},
  \code{\link{is_single_string}},
  \code{\link{is_string_list}}, \code{\link{is_time_unit}},
  \code{\link{is_unit}}, \code{\link{rm_units}},
  \code{\link{single_na}}, \code{\link{to_range}}
}
\concept{helper functions}
