% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_single_numeric}
\alias{is_single_numeric}
\title{Checks if Input is a Single Bumeric Value}
\usage{
is_single_numeric(input)
}
\arguments{
\item{input}{The input that should be tested}
}
\value{
True if the input is a single numeric value. Lists with one numeric value will return False.
}
\description{
Checks if Input is a Single Bumeric Value
}
\seealso{
Other helper functions: \code{\link{has_units}},
  \code{\link{is_single_logical}},
  \code{\link{is_single_string}},
  \code{\link{is_string_list}}, \code{\link{is_unit}}
}
\concept{helper functions}
