% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_single_string}
\alias{is_single_string}
\title{Checks if Input is a Single String}
\usage{
is_single_string(input, can.be.empty = T, trim.check = T)
}
\arguments{
\item{input}{The input that should be tested}

\item{can.be.empty}{If the input (if string) is allowed to be empty.}

\item{trim.check}{If the input (if string) should be trimmed before the check.}
}
\value{
True if the input is a single string of vector with a single string.
Lists with one string element will return False.
}
\description{
Checks if Input is a Single String
}
\seealso{
Other helper functions: \code{\link{has_units}},
  \code{\link{is_single_logical}},
  \code{\link{is_single_numeric}},
  \code{\link{is_string_list}}, \code{\link{is_unit}}
}
\concept{helper functions}
