% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_string_list}
\alias{is_string_list}
\title{Tests if Input is a String or List-like Object of Strings}
\usage{
is_string_list(vec, string.like = F, can.be.empty = F)
}
\arguments{
\item{vec}{The input object.}

\item{string.like}{If input is allowed to be convertible to string for the test.}

\item{can.be.empty}{If empty strings are allowed.

Returns `TRUE` for empty lists. For lists with NA-values this function will return `FALSE`.}
}
\value{
`TRUE`, if the input is a string or list-like object of strings, else `FALSE`.
}
\description{
Tests if Input is a String or List-like Object of Strings
}
\seealso{
Other helper functions: \code{\link{any_na}},
  \code{\link{any_null}}, \code{\link{convert_values}},
  \code{\link{has_dose}}, \code{\link{has_units}},
  \code{\link{is_dose_unit}},
  \code{\link{is_numeric_list}},
  \code{\link{is_single_logical}},
  \code{\link{is_single_numeric}},
  \code{\link{is_single_string}},
  \code{\link{is_time_unit}}, \code{\link{is_unit}},
  \code{\link{rm_units}}, \code{\link{single_na}},
  \code{\link{to_range}}
}
\concept{helper functions}
