% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_time_unit}
\alias{is_time_unit}
\title{Tests for a Valid Time Unit String}
\usage{
is_time_unit(str)
}
\arguments{
\item{str}{The string that should be tested.}
}
\value{
`TRUE`, if a valid time unit is provided, else `FALSE`.
}
\description{
This function will only work correctly for lower case strings (e.g. h and not H).
For non-strings, NA or NULL this function will return False.
}
\examples{

is_time_unit("") # F
is_time_unit(NA) # F
is_time_unit("h") # T
is_time_unit("sec") # T
is_time_unit("weeks") # T

}
\seealso{
Other helper functions: \code{\link{any_na}},
  \code{\link{any_null}}, \code{\link{convert_values}},
  \code{\link{has_dose}}, \code{\link{has_units}},
  \code{\link{is_dose_unit}},
  \code{\link{is_numeric_list}},
  \code{\link{is_single_logical}},
  \code{\link{is_single_numeric}},
  \code{\link{is_single_string}},
  \code{\link{is_string_list}}, \code{\link{is_unit}},
  \code{\link{rm_units}}, \code{\link{single_na}},
  \code{\link{to_range}}
}
\concept{helper functions}
