% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{is_unique_molecule_list}
\alias{is_unique_molecule_list}
\title{Checks if a Molecule List has Unique Ids}
\usage{
is_unique_molecule_list(mol.list)
}
\arguments{
\item{mol.list}{The molecule list that should be tested.}
}
\value{
True, if the molecule list is unique, else False.
}
\description{
Checks if a Molecule List has Unique Ids
}
\seealso{
Other molecule functions: \code{\link{has_molecules}},
  \code{\link{is_molecule_list}},
  \code{\link{is_molecule}}, \code{\link{match_mol}},
  \code{\link{match_pvb}},
  \code{\link{merge_molecule_lists}},
  \code{\link{molecule_from_ids}},
  \code{\link{molecule_ids}}, \code{\link{molecule_list}},
  \code{\link{molecule}}, \code{\link{print.molecule}}
}
\concept{molecule functions}
