% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_unit}
\alias{is_unit}
\title{Checks if a String is a Valid Unit}
\usage{
is_unit(str)
}
\arguments{
\item{str}{The input string.}
}
\value{
True if a valid unit, else False.
}
\description{
Checks if a String is a Valid Unit
}
\seealso{
Other helper functions: \code{\link{has_units}},
  \code{\link{is_single_logical}},
  \code{\link{is_single_numeric}},
  \code{\link{is_single_string}},
  \code{\link{is_string_list}}
}
\concept{helper functions}
