% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{match_mol}
\alias{match_mol}
\title{Escapes a Name in a Reg-exp Friendly Manner to Match a Molecule in a Compartment.}
\usage{
match_mol(name)
}
\arguments{
\item{name}{A non-empty string that should be escaped.}
}
\value{
An escapted name.
}
\description{
Escapes a Name in a Reg-exp Friendly Manner to Match a Molecule in a Compartment.
}
\seealso{
Other molecule functions: \code{\link{has_molecules}},
  \code{\link{is_molecule_list}},
  \code{\link{is_molecule}},
  \code{\link{is_unique_molecule_list}},
  \code{\link{match_pvb}},
  \code{\link{merge_molecule_lists}},
  \code{\link{molecule_from_ids}},
  \code{\link{molecule_ids}}, \code{\link{molecule_list}},
  \code{\link{molecule}}, \code{\link{print.molecule}}
}
\concept{molecule functions}
