% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{match_pvb}
\alias{match_pvb}
\title{Returns a Match String for Peripheral Venous Blood}
\usage{
match_pvb()
}
\value{
An match string.
}
\description{
Returns a Match String for Peripheral Venous Blood
}
\seealso{
Other molecule functions: \code{\link{has_molecules}},
  \code{\link{is_molecule_list}},
  \code{\link{is_molecule}},
  \code{\link{is_unique_molecule_list}},
  \code{\link{match_mol}},
  \code{\link{merge_molecule_lists}},
  \code{\link{molecule_from_ids}},
  \code{\link{molecule_ids}}, \code{\link{molecule_list}},
  \code{\link{molecule}}, \code{\link{print.molecule}}
}
\concept{molecule functions}
