% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{merge_molecule_lists}
\alias{merge_molecule_lists}
\title{Merges two Molecule Lists}
\usage{
merge_molecule_lists(list.a, list.b)
}
\arguments{
\item{list.a}{The first molecule list (must be non-empty).}

\item{list.b}{The second molecule list (must be non-empty).}
}
\value{
A merged molecule list with unique ids.
}
\description{
Trick: You can make a molecule list unique by merging the list with itself.
}
\examples{

mol_1 <- molecule("Mol 1", MW = 100)
mol_2 <- molecule("Mol 2", MW = 200)
mol_list <- molecule_list(mol_1, mol_2)
merge_molecule_lists(mol_list, mol_list) # returns mol_list

}
\seealso{
Other molecule functions: \code{\link{has_molecules}},
  \code{\link{is_molecule_list}},
  \code{\link{is_molecule}},
  \code{\link{is_unique_molecule_list}},
  \code{\link{match_mol}}, \code{\link{match_pvb}},
  \code{\link{molecule_from_ids}},
  \code{\link{molecule_ids}}, \code{\link{molecule_list}},
  \code{\link{molecule}}, \code{\link{print.molecule}}
}
\concept{molecule functions}
