% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{molecule_ids}
\alias{molecule_ids}
\title{Returns a vector of Molecule Ids from a Molecule List}
\usage{
molecule_ids(mol.list)
}
\arguments{
\item{mol.list}{A non-empty molecule list}
}
\value{
A vector of molecule Ids
}
\description{
Returns a vector of Molecule Ids from a Molecule List
}
\examples{

mol_1 <- molecule("Mol 1", MW = 100)
mol_2 <- molecule("Mol 2", MW = 200)
mol_list <- molecule_list(mol_1, mol_2)
ids <- molecule_ids(mol_list)

}
\seealso{
Other molecule functions: \code{\link{has_molecules}},
  \code{\link{is_molecule_list}},
  \code{\link{is_molecule}},
  \code{\link{is_unique_molecule_list}},
  \code{\link{merge_molecule_lists}},
  \code{\link{molecule_from_ids}},
  \code{\link{molecule_list}}, \code{\link{molecule}},
  \code{\link{print.molecule}}
}
\concept{molecule functions}
