% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{molecule_list}
\alias{molecule_list}
\title{Creates a List of Molecules}
\usage{
molecule_list(...)
}
\arguments{
\item{...}{Must be a non-empty list of molecules.}
}
\value{
A list of molecules that are named by id
}
\description{
Creates a List of Molecules
}
\examples{

mol_1 <- molecule("Mol 1", MW = 100)
mol_2 <- molecule("Mol 2", MW = 200)
mol_list <- molecule_list(mol_1, mol_2)
is_molecule_list(mol_list) # is TRUE

}
\seealso{
Other molecule functions: \code{\link{has_molecules}},
  \code{\link{is_molecule_list}},
  \code{\link{is_molecule}},
  \code{\link{is_unique_molecule_list}},
  \code{\link{merge_molecule_lists}},
  \code{\link{molecule_from_ids}},
  \code{\link{molecule_ids}}, \code{\link{molecule}},
  \code{\link{print.molecule}}
}
\concept{molecule functions}
