% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_master}
\alias{parse_master}
\title{Title}
\usage{
parse_master(file_or_data, id, molecules, entry.sep = ",", ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
\examples{

}
\seealso{
Other parser functions: \code{\link{excel_sheets}},
  \code{\link{extract_unit}}, \code{\link{file_exists}},
  \code{\link{file_type}}, \code{\link{ident_time_column}},
  \code{\link{read_file_table}}
}
\concept{parser functions}
