% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{print.molecule}
\alias{print.molecule}
\title{Prints Information about a Molecule}
\usage{
\method{print}{molecule}(x, details = F, ...)
}
\arguments{
\item{x}{The molecule object.}

\item{details}{True if all details should be printed, else False.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
Invisible copy of the object \code{x}.
}
\description{
Prints Information about a Molecule
}
\examples{
a <- molecule("Nicotine")
a # prints a without details
print(a, details = TRUE) # prints a with details

b <- molecule("THC")
mol_list <- list(a, b)
print(mol_list, details = TRUE) # prints molecule list with details

}
\seealso{
Other molecule functions: \code{\link{has_molecules}},
  \code{\link{is_molecule_list}},
  \code{\link{is_molecule}},
  \code{\link{is_unique_molecule_list}},
  \code{\link{merge_molecule_lists}},
  \code{\link{molecule_from_ids}},
  \code{\link{molecule_ids}}, \code{\link{molecule_list}},
  \code{\link{molecule}}
}
\concept{molecule functions}
