% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{q05}
\alias{q05}
\title{Calculate the 5th Percentile of a Numeric Vector}
\usage{
q05(values)
}
\arguments{
\item{values}{A numeric vector containing the values for which the 5th percentile is to be calculated.}
}
\value{
A single numeric value representing the 5th percentile of the input vector.
}
\description{
This function computes the 5th percentile (0.05 quantile) of a given numeric vector.
}
\examples{
q05(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)) # Returns 1.45
q05(c(10, 20, 30, 40, 50)) # Returns 11
}
