% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{q16}
\alias{q16}
\title{Calculate the 16th Percentile}
\usage{
q16(values)
}
\arguments{
\item{values}{A numeric vector containing the values for which the 16th percentile is to be calculated.}
}
\value{
A single numeric value representing the 16th percentile of the input vector.
}
\description{
This function computes the 16th percentile of a given numeric vector.
}
\examples{
q16(c(1, 2, 3, 4, 5)) # Returns the 16th percentile of the vector
}
