% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{q84}
\alias{q84}
\title{Calculate the 84th Percentile of a Numeric Vector}
\usage{
q84(values)
}
\arguments{
\item{values}{A numeric vector containing the values for which the 84th percentile is to be calculated.}
}
\value{
A single numeric value representing the 84th percentile of the input vector.
}
\description{
This function computes the 84th percentile of a given numeric vector.
}
\examples{
q84(c(1, 2, 3, 4, 5)) # Returns the 84th percentile
q84(rnorm(100)) # Returns the 84th percentile of 100 random normal values
}
