% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{q95}
\alias{q95}
\title{Calculate the 95th Percentile}
\usage{
q95(values)
}
\arguments{
\item{values}{A numeric vector containing the values for which the 95th percentile is to be calculated.}
}
\value{
A single numeric value representing the 95th percentile of the input vector.
}
\description{
This function computes the 95th percentile of a given numeric vector.
}
\examples{
q95(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)) # Returns 9.55
q95(c(10, 20, 30, 40, 50)) # Returns 48
}
