% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{read_file_table}
\alias{read_file_table}
\title{Title}
\usage{
read_file_table(file, file.type = c("auto", "excel", "csv", "tsv"),
  excel.sheet = 1, csv.sep = if (file.type == "tsv") "\\t" else ",",
  csv.encoding = "guess", csv.dec = ".", csv.group = ",",
  csv.comment = "", skip.lines = 0)
}
\arguments{
\item{skip.lines}{}
}
\description{
Title
}
\seealso{
Other parser functions: \code{\link{excel_sheets}},
  \code{\link{extract_unit}}, \code{\link{file_exists}},
  \code{\link{file_type}}, \code{\link{ident_time_column}},
  \code{\link{parse_master}}
}
\concept{parser functions}
