% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{rm_units}
\alias{rm_units}
\title{Removes Units from an Object}
\usage{
rm_units(x)
}
\arguments{
\item{x}{The input object.}
}
\value{
`x` without units attached.
}
\description{
If the input does not have units attached it is returned.
}
\seealso{
Other helper functions: \code{\link{any_na}},
  \code{\link{any_null}}, \code{\link{convert_values}},
  \code{\link{has_dose}}, \code{\link{has_units}},
  \code{\link{is_dose_unit}},
  \code{\link{is_numeric_list}},
  \code{\link{is_single_logical}},
  \code{\link{is_single_numeric}},
  \code{\link{is_single_string}},
  \code{\link{is_string_list}}, \code{\link{is_time_unit}},
  \code{\link{is_unit}}, \code{\link{single_na}},
  \code{\link{to_range}}
}
\concept{helper functions}
