% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{single_na}
\alias{single_na}
\title{Convinience Helper to Test for single `NA` value}
\usage{
single_na(x)
}
\arguments{
\item{x}{The test value.}
}
\value{
`TRUE` if first value in `x` is `NA` and `length(x) == 1`.
}
\description{
Convinience Helper to Test for single `NA` value
}
\seealso{
Other helper functions: \code{\link{any_na}},
  \code{\link{any_null}}, \code{\link{convert_values}},
  \code{\link{has_dose}}, \code{\link{has_units}},
  \code{\link{is_dose_unit}},
  \code{\link{is_numeric_list}},
  \code{\link{is_single_logical}},
  \code{\link{is_single_numeric}},
  \code{\link{is_single_string}},
  \code{\link{is_string_list}}, \code{\link{is_time_unit}},
  \code{\link{is_unit}}, \code{\link{rm_units}},
  \code{\link{to_range}}
}
\concept{helper functions}
