% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{std.dev.max}
\alias{std.dev.max}
\title{Calculate the maximum Standard Deviation}
\usage{
std.dev.max(values)
}
\arguments{
\item{values}{A numeric vector containing the values to calculate the mean
and standard deviation.}
}
\value{
A numeric value representing the mean plus the standard deviation.
}
\description{
This function computes the maximum value by adding the standard deviation
of a numeric vector to its mean.
}
\examples{
values <- c(1, 2, 3, 4, 5)
std.dev.max(values) # Returns mean(values) + sd(values)

}
