% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{std.dev.min}
\alias{std.dev.min}
\title{Calculate the Minimum Standard Deviation}
\usage{
std.dev.min(values)
}
\arguments{
\item{values}{A numeric vector. The input values for which the mean and standard
deviation will be calculated.}
}
\value{
A numeric value representing the mean minus the standard deviation of
the input vector.
}
\description{
This function computes the difference between the mean and the standard deviation
of a numeric vector of values.
}
\examples{
values <- c(1, 2, 3, 4, 5)
std.dev.min(values)

}
