% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{[.master}
\alias{[.master}
\title{Extracts One or More entries (Row Entries) from a Master Object}
\usage{
\method{[}{master}(x, ...)
}
\arguments{
\item{x}{The master object.}

\item{...}{One or more (`vector`) numerical idexes (row numbers) of the master object data.

`...` behaves like `vector[idx]`. E.g. `idx = -1` is allowed and does what the user would
  expect.}
}
\value{
A list of entries and deserialized molecules, ids and ranges.
 Entries have class `masterEntry`. If only one object is fetched only one `masterEntry` object
 is returned.
}
\description{
Extracts One or More entries (Row Entries) from a Master Object
}
\examples{

mol <- molecule("Mol 1", MW = 100)
m1 <- master("test", mol)
m1 <- add_master_entry(m1, "Pop 1", pop.molecules = mol)
m1 <- add_master_entry(m1, "Pop 2", pop.molecules = mol)

m_filtered <- m1[1] # one entry (not a list of entries)
m_filtered_2 <- m1[1,2] # two entries

}
\seealso{
Other master functions: \code{\link{add_master_entry}},
  \code{\link{bind_master}}, \code{\link{entry.default}},
  \code{\link{entry.master}}, \code{\link{entry}},
  \code{\link{filter_master}},
  \code{\link{groups.default}},
  \code{\link{groups.master}}, \code{\link{groups}},
  \code{\link{is_masterEntry_list}},
  \code{\link{is_masterEntry}}, \code{\link{is_master}},
  \code{\link{master}}, \code{\link{print.master}}
}
\concept{master functions}
