% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{to_range}
\alias{to_range}
\title{Converts two Values and a Unit String to a Valid Range}
\usage{
to_range(min, max, unit.str)
}
\arguments{
\item{min}{The minumum value (will be converted by `as.numeric`).}

\item{max}{The maximum value (will be converted by `as.numeric`)}

\item{unit.str}{A unit string (e.g. mg).}
}
\value{
A vector with `length` 2 that has a unit attached. If `min` and `max` are both `NA` no unit
will be attached.

If `min` and `max` are both `NA` `unit.str` is ignored and will not be evaluated.
}
\description{
Converts two Values and a Unit String to a Valid Range
}
\examples{

to_range(NA, NA, NA) # will return c(NA, NA)
to_range(NA, 1, "mg") # will return c(NA, 1) with attached unit "mg"
to_range(NA, "1", "mg") # will return c(NA, 1) with attached unit "mg"
# to_range(1, NA, NA) # error

}
\seealso{
Other helper functions: \code{\link{any_na}},
  \code{\link{any_null}}, \code{\link{convert_values}},
  \code{\link{has_dose}}, \code{\link{has_units}},
  \code{\link{is_dose_unit}},
  \code{\link{is_numeric_list}},
  \code{\link{is_single_logical}},
  \code{\link{is_single_numeric}},
  \code{\link{is_single_string}},
  \code{\link{is_string_list}}, \code{\link{is_time_unit}},
  \code{\link{is_unit}}, \code{\link{rm_units}},
  \code{\link{single_na}}
}
\concept{helper functions}
