% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{add_master_entry}
\alias{add_master_entry}
\title{Adds a New Entry to an Master Object}
\usage{
add_master_entry(master.obj, pop.name = NA, pop.molecules = NA,
  sim.name = NA, sim.molecules = NA, obs.ids = NA, groups = NA,
  plot.header = NA, plot.x.range = NA, plot.y.range = NA)
}
\arguments{
\item{master.obj}{The master object (must be one valid master object).}

\item{pop.name}{The name of the population simulation (must be NA or a single non-empty string).}

\item{pop.molecules}{One or more (molecule list) molecules that are associated to the population simulation. Can be NA if \code{pop.name} is NA.}

\item{sim.name}{The name of the individual simulation (must be NA or a single non-empty string).}

\item{sim.molecules}{One or more (molecule list) molecules that are associated to the individual simulation. Can be NA if \code{sim.name} is NA.}

\item{obs.ids}{The associated observed data ids. Can be NA or one/more string-convertibles (e.g. strings or numbers). The Ids are converted to strings.}

\item{groups}{Group entries for the simulations. Must be NA or a list-like object with string-convertibles or NA (if not set for a group).
The number of elements cannot be larger than the number of groups of the master object.}

\item{plot.header}{The name of the plottings header for the simulations. Can be NA or a (empty) single string.}

\item{plot.x.range}{The plotting range for the x-axis. Can be NA (no range). If not NA, units must be associated with the numeric vector.}

\item{plot.y.range}{The plotting range for the y-axis. Can be NA (no range). If not NA, units must be associated with the numeric vector.

A note on the plot.X/Y.range options: If only one parameter (with unit) is provided the range is implicitly set to \code{c(value, NA)} (no upper limit).
If NA is set (one value; no limits) no unit must be set. If only one limit (upper or lower) should be set the complimentary direction must be set to NA.

A note on molecules: Molecules must be present in the master object (when master was created). Unknown molecules will raise an error.}
}
\value{
The master object with added entry.
}
\description{
Adds a New Entry to an Master Object
}
\examples{

mol_1 <- molecule("Foo", MW = 100)
mol_2 <- molecule("Bar", MW = 200)
m <- master("Test", molecule_list(mol_1, mol_2), groups = 2) # master with 2 associated molecules

m <- add_master_entry(m, pop.name = "Pop", pop.molecules = mol_1) # ok
m <- add_master_entry(m, pop.name = "Pop 2", pop.molecules = molecule_list(mol_1, mol_2)) # ok
# m <- add_master_entry(m, pop.name = "Pop 2", pop.molecules = NA) # error
# m <- add_master_entry(m, pop.name = "Pop 2", pop.molecules = molecule("Unknown", MW = 10)) # error

m <- add_master_entry(m, pop.name = "Pop", pop.molecules = mol_1, groups = c("Foo", NA)) # ok only Group_1 is set
m <- add_master_entry(m, pop.name = "Pop", pop.molecules = mol_1, groups = c("Foo", "Boo")) # ok both groups set
# m <- add_master_entry(m, pop.name = "Pop", pop.molecules = mol_1, groups = c("Foo", "Boo", "Not in master")) # error

m <- add_master_entry(m, plot.x.range = NA) # ok -> no limit
m <- add_master_entry(m, plot.x.range = units::as_units(c(1,2), "h")) # ok upper and lower limit
m <- add_master_entry(m, plot.x.range = units::as_units(c(NA, 2), "h")) # ok only upper limit
m <- add_master_entry(m, plot.x.range = units::as_units(c(1, NA), "h")) # ok only lower limit

# m <- add_master_entry(m, plot.x.range = c(1,2)) # error -> no units provided

}
\seealso{
Other master functions: \code{\link{bind_master}},
  \code{\link{filter_master}}, \code{\link{is_master}},
  \code{\link{master}}, \code{\link{print.master}}
}
\concept{master functions}
