% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{is_string_list}
\alias{is_string_list}
\title{Tests if Input is a String or List-like Object of Strings}
\usage{
is_string_list(vec, string.like = F, can.be.empty = F)
}
\arguments{
\item{vec}{The input object.}

\item{string.like}{If input is allowed to be convertible to string for the test.}

\item{can.be.empty}{If empty strings are allowed.}
}
\value{
True, if the input is a string or list-like object of strings, else False.
}
\description{
Tests if Input is a String or List-like Object of Strings
}
\seealso{
Other helper functions: \code{\link{has_units}},
  \code{\link{is_single_logical}},
  \code{\link{is_single_numeric}},
  \code{\link{is_single_string}}, \code{\link{is_unit}}
}
\concept{helper functions}
