% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{master}
\alias{master}
\title{Creates of Master Object}
\usage{
master(id, molecules, reference = NA, groups = 0)
}
\arguments{
\item{id}{The id of the master object (must be a non-empty string).}

\item{molecules}{A molecule list or single molecule that should at least contain all molecules that are added later on. Must be non-empty.

Adding entries to a master object should only be done via \code{add_master_entry}.}

\item{reference}{The reference of the master object (optional).}

\item{groups}{The number of Groups per master entry that should be created (must be numberic and at least be 0).}
}
\value{
A master object.
}
\description{
The reference should typically store the filename if the master object is parsed from a file.
}
\examples{

m1 <- master("test")
m2 <- master("test 2", "my ref", groups = 3)

}
\seealso{
Other master functions: \code{\link{add_master_entry}},
  \code{\link{bind_master}}, \code{\link{filter_master}},
  \code{\link{is_master}}, \code{\link{print.master}}
}
\concept{master functions}
