% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{molecule}
\alias{molecule}
\title{Creates a Molecule}
\usage{
molecule(name, display.name = name, id = name, container = NA,
  column.matcher = c(name), pubchem.id = NA, MW = NA,
  is.fraction = F, fixed.unit = NA, ylab = "Plasma Concentration",
  col = "black", pch = 19, lty = 1)
}
\arguments{
\item{name}{The name of the molecule (must be a single non-empty string).}

\item{display.name}{The name that should be displayed in plots (e.g. in the legend). Must be a single non-empty string.}

\item{id}{A unique id for the Molecule. Must be a single value that will be converted to string.}

\item{container}{A single non-empty string or NA that defines the model container the molecule is simulated.}

\item{column.matcher}{Strings that match the Molecule for the parser functions (AND relationship).
Must be a vector/single value of non-empty strings.}

\item{pubchem.id}{The pubchem identifier. Must be a single numeric value or NA.}

\item{MW}{The molecular weight of the Molecule. Must be a single numeric value or NA.}

\item{is.fraction}{Single logical value that indicates if the Molecule represents a fraction (e.g. fraction excreted to urin).}

\item{fixed.unit}{Indicates if the Molecule has a fixed unit (always the same unit). If not a single unit is used NA must be set.}

\item{ylab}{A label for the y-axis if Profiles for a Molecule are plotted. Must be a single non-empty string.}

\item{col}{A color that should be used for plotting. Must be a single value (NA is allowed).}

\item{pch}{A point type that should be used for plotting. Must be a single value (NA is allowed).}

\item{lty}{A line type that should be used for plotting. Must be a single value (NA is allowed).}
}
\description{
A molecule is a description of a measured entity from a PKSim/Mobi container.
It can be a PK descriptor (e.g. concentrations) or PD descriptor (e.g. heart rate).
}
\details{
If MW is not set, the name or (if set) the pubchem identifier will be used to fetch the molecular weight
from the Pubchem database. If this fails the molecule is not created. If the fetching is done via name
(no pubchem identifier is provided) the pubchem idenfier is fetched as well.

For PD Molecules a dummy value for MW should be set (e.g. 0) to avoid the automatic fetching from Pubchem.
When it comes to matchers, special charaters (e.g. "|") must be regular expression safe (escaped).
We povide special matcher functions that allows the easy creation often used matching expressions.
}
\seealso{
Other molecule functions: \code{\link{has_molecules}},
  \code{\link{is_molecule_list}},
  \code{\link{is_molecule}},
  \code{\link{is_unique_molecule_list}},
  \code{\link{match_mol}}, \code{\link{match_pvb}},
  \code{\link{merge_molecule_lists}},
  \code{\link{molecule_from_ids}},
  \code{\link{molecule_ids}}, \code{\link{molecule_list}},
  \code{\link{print.molecule}}
}
\concept{molecule functions}
