% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{print.master}
\alias{print.master}
\title{Prints Information about a Master Object}
\usage{
\method{print}{master}(x, details = F, max = NA)
}
\arguments{
\item{x}{The Master object.}

\item{details}{True if all details should be printed, else False.}

\item{max}{Maximum number of data rows to print. If \code{max = NA} (default) the number of entries that are printed
is \code{getOption("max.print")/ncol(data)} for detailed print and the number of rows printed is 10 for non-detailed print.}
}
\value{
Invisible copy of the object \code{x}.
}
\description{
Prints Information about a Master Object
}
\examples{
a <- master("Master 1")
a # prints a without details
print(a, details = T) # prints a with details
print(a, max = 5) # prints a without details and maximum of 5 data rows

}
\seealso{
Other master functions: \code{\link{add_master_entry}},
  \code{\link{bind_master}}, \code{\link{filter_master}},
  \code{\link{is_master}}, \code{\link{master}}
}
\concept{master functions}
